Add data pathway for multi-sens scenario data: - Add boilerplate 3d plots - Add data retrieval methods - Move data retrieval methods to scribe class
